added user table and adapter for users
TODO:
Change DB datatype for OWNER from VARCHAR to NUMBER in the TRADE table.

Currently the relation between trade and user is in the user NAME.
However this is very bad and dumb in case the user wants to change their
name at some point.

should be an easy fix.